証明の登録（Structの場合）ざっくり実装

+++ Models/ProofInferenceArgument.cs
@@ -34,5 +34,5 @@ public class ProofInferenceArgument
   /// <summary>
   /// 論理式構成ID
   /// </summary>
-  public int? FormulaStructId { get; set; }  
+  public long? FormulaStructId { get; set; }  
 }
